+++ src/messages.ts
@@ -4,7 +4,8 @@ export const ERROR_MESSAGES = {
     'noActiveFile': "No document is opened in the VSCode editor.",
     'nonPythonFile': "Cannot infer type annotations for non-Python code files.",
     'emptyFile': "Cannot infer type annotations for empty files.",
-    'emptyPayload': "The received response was empty."
+    'emptyPayload': "The received response was empty.",
+    'connectionError': "Could not connect to the server."
 };
 
 export const TELEMETRY_REQUEST_MESSAGE = "Would you like to share accepted " +