+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to the [Type4Py's VSCode extension](https://github.com/saltu
 
 ## [Unreleased]
 ### Added
+- Include the name of the extension in the right side of predicted types.
+- A naive support for predicting types in the case of variables redeclaration in the scope of functions (Addresses [#7](https://github.com/saltudelft/type4py-vscode-ext/issues/7)).
+- A status bar at the bottom left of VSCode, which shows no. of in-progress and completed type prediction requests.
+- An output channel for the extension to show its log messages instead of showing info windows.
+
+## [0.1.4] - 2021-07-21
+### Added
 - Detecting `ExtensionMode` for testing and development using the local server (**ONLY FOR THE EXTENSION'S DEVELOPERS**).
 - Submitting the canceled/rejected type predictions based on the user's consent.
 - Sending the hash of files' absolute path, the extension's version, and activation ID along with the prediction request (**ONLY FOR RESEARCH PURPOSE**).